requirements.txt
alembic==1.13.3
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-CKEditor==1.0.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jaraco.text==4.0.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.5.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==75.1.0
SQLAlchemy==2.0.35
tomli==2.0.1
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
WTForms==3.1.2
zipp==3.20.2
alembic==1.13.3
autocommand==2.2.2
backports.tarfile==1.2.0
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-CKEditor==1.0.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jaraco.text==4.0.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.5.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
tomli==2.0.1
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.4
wheel==0.44.0
WTForms==3.1.2
zipp==3.20.2
